feat: Add functionality to create an icon based on user input
This commit adds a new HTML file, `html13_randomSpan.html`, which allows users to enter text in an input field and create an icon based on that input. The icon is displayed using the Material Icons Sharp font. When the user clicks the "Create Icon" button, the text from the input field is set as the inner HTML of a `<span>` element with the id "icon".

The purpose of this change is to enhance the user experience by providing a way for users to dynamically create icons based on their input.